fix a few warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@20602 f0d54148-0727-0410-94bb-9a71ac55c965

+++ ogr/ogrsf_frmts/kml/kmlnode.cpp
@@ -652,7 +652,7 @@ Feature* KMLNode::getFeature(std::size_t nNum, int& nLastAsked, int &nLastCount)
     if(nNum >= this->getNumFeatures())
         return NULL;
 
-    if (nLastAsked + 1 != nNum)
+    if (nLastAsked + 1 != (int)nNum)
     {
         nCount = 0;
         nCountP = 0;
